Re-structuring + docstrings + typing
